add database seeder and tests

+++ src/main/java/br/com/fiap/medieval_trade_api/model/Item.java
@@ -1,14 +1,7 @@
 package br.com.fiap.medieval_trade_api.model;
 
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
@@ -16,14 +9,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-
 public class Item {
-    
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -36,9 +30,9 @@ public class Item {
 
     @Enumerated(EnumType.STRING)
     private ItemRarity rarity;
-    
+
     @Positive
-    private Integer price;
+    private BigDecimal price;
 
     @ManyToOne
     @JoinColumn(name = "owner_id")
